Arduino code implementation
The code to run in the arduino is implemented.

The code itself makes a delta between the values of pairs of LDR and
map them into a delta of degrees to rotate the servo, in each 15 ms. As
the configuration goes, the movement is decoupled in a way that the horizontal rotation of the solar panel is dictated by the values of LDR1 and LDR3, the lower and upper sensors of the panel and perpendicular to the axis of rotation of the motor, and the vertical rotation of the solar panel is dictaated by the values of LDR2 and LD4, the lateral sensors, with the former being at the right and the latter at the left.

After the calculations, the current position for each servo is updated
and then waits 15ms for the next reading of the values.

As for correctness, this code wasn't tested and there may appear some
values equal to zero, but it's possible after some iterations of this
project to simplify and solve all the problems without a sweat.